rke2-uninstall.sh enhancements
Fix a couple gaps/oversights in uninstall. First don't remove /etc/rancher/rke2 unconditionally so that if a user created config.yaml, or any other file is present we don't remove user data that may not be able to be recreated. Aka a hand crafted config.yaml file that isn't backed up shouldn't be removed.

Secondarily, check if /var/lib/rancher is a mountpoint or not, if it is do not try to remove it with rm.